+++ locallib.php
@@ -183,9 +183,9 @@ function theme_boost_campus_get_imageareacontent() {
 
         // Only continue processing if there are files in the filearea.
         if (!empty($files)) {
-            // Get the content from the setting imageareaitemslink and explode it to an array by the delimiter "new line".
+            // Get the content from the setting imageareaitemsattributes and explode it to an array by the delimiter "new line".
             // The string contains: the image identifier (uploaded file name) and the corresponding link URL.
-            $lines = explode("\n", get_config('theme_boost_campus', 'imageareaitemslink'));
+            $lines = explode("\n", get_config('theme_boost_campus', 'imageareaitemsattributes'));
             // Parse item settings.
             foreach ($lines as $line) {
                 $line = trim($line);
@@ -194,41 +194,64 @@ function theme_boost_campus_get_imageareacontent() {
                     // Create an array with a dummy entry because the function array_key_exists need a
                     // not empty array for parameter 2.
                     $links = array('foo');
+                    $alttexts = array('bar');
                     continue;
                 } else {
                     $settings = explode("|", $line);
-                    // Check if both parameters are set.
-                    if (!empty($settings[1])) {
-                        // The name of the image is the key for the URL that will be set.
-                        $links[$settings[0]] = $settings[1];
+                    // Check if parameter 2 or 3 is set.
+                    if (!empty($settings[1] || !empty($settings[2]))) {
+                        foreach ($settings as $i => $setting) {
+                            $setting = trim($setting);
+                            if (!empty($setting)) {
+                                switch ($i) {
+                                    // Check for the first param: link.
+                                    case 1:
+                                        // The name of the image is the key for the URL that will be set.
+                                        $links[$settings[0]] = $settings[1];
+                                        break;
+                                    // Check for the second param: alt text.
+                                    case 2:
+                                        // The name of the image is the key for the alt text that will be set.
+                                        $alttexts[$settings[0]] = $settings[2];
+                                        break;
+                                }
+                            }
+                        }
                     }
                 }
             }
+            // Initialize the array which holds the data which is later stored in the cache.
+            $imageareacache = [];
             // Traverse the files.
             foreach ($files as $file) {
                 // Get the Moodle url for each file.
                 $url = moodle_url::make_pluginfile_url($file->get_contextid(), $file->get_component(), $file->get_filearea(),
-                    $file->get_itemid(), $file->get_filepath(), $file->get_filename());
+                        $file->get_itemid(), $file->get_filepath(), $file->get_filename());
                 // Get the path to the file.
                 $filepath = $url->get_path();
                 // Get the filename.
                 $filename = $file->get_filename();
-                // If filename and key value from the imageareaitemslink setting entry match.
+                // If filename and link value from the imageareaitemsattributes setting entry match.
                 if (array_key_exists($filename, $links)) {
-                    // Set the file and the corresponding link.
-                    $imageareacache[] = array('filepath' => $filepath, 'linkpath' => $links[$filename]);
-                    // Fill the cache.
-                    $themeboostcampuscache->set('imageareadata', $imageareacache);
-                } else { // Just add the file without a link.
-                    $imageareacache[] = array('filepath' => $filepath);
-                    // Fill the cache.
-                    $themeboostcampuscache->set('imageareadata', $imageareacache);
+                    $linkpath = $links[$filename];
+                } else {
+                    $linkpath = "";
                 }
+                // If filename and alt text value from the imageareaitemsattributes setting entry match.
+                if (array_key_exists($filename, $alttexts)) {
+                    $alttext = $alttexts[$filename];
+                } else {
+                    $alttext = "";
+                }
+                // Add the file
+                $imageareacache[] = array('filepath' => $filepath, 'linkpath' => $linkpath, 'alttext' => $alttext);
             }
             // Sort array alphabetically ascending to the key "filepath".
             usort($imageareacache, function($a, $b) {
                 return strcmp($a["filepath"], $b["filepath"]);
             });
+            // Fill the cache.
+            $themeboostcampuscache->set('imageareadata', $imageareacache);
             return $imageareacache;
         } else { // If no images are uploaded, then cache an empty array.
             return $themeboostcampuscache->set('imageareadata', array());